rename param files
