+++ zenu-matrix/src/nn/rnn/mod.rs
@@ -1,7 +1,7 @@
 mod descriptor;
-mod params;
 mod rnn;
+mod rnn_params;
 mod test;
 
 pub use descriptor::*;
-pub use params::*;
+pub use rnn_params::*;
